+++ tmf630-toolkit-jsonb/src/test/java/org/opentmf/query/tmf630/jsonb/it/SplitOrderItem.java
@@ -1,0 +1,28 @@
+package org.opentmf.query.tmf630.jsonb.it;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class SplitOrderItem {
+
+  private String id;
+  private String state;
+
+  public String getId() {
+    return id;
+  }
+
+  public void setId(String id) {
+    this.id = id;
+  }
+
+  public String getState() {
+    return state;
+  }
+
+  public void setState(String state) {
+    this.state = state;
+  }
+}