+++ tmf630-toolkit-jsonb/src/test/java/org/opentmf/query/tmf630/jsonb/it/SplitOrderRowRepository.java
@@ -1,0 +1,5 @@
+package org.opentmf.query.tmf630.jsonb.it;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SplitOrderRowRepository extends JpaRepository<SplitOrderRow, String> {}